Finish orbit token bridge deployer scripts
